requirements.txt
requests
python-dotenv
web3
gspread
google-auth
solana
base58
solders